// src/pages/System/types/SystemGetApi.type.ts
export interface ConfigurationGetResponse {
  active: boolean;
  /**
   * ID
   */
  id: string;
  /**
   * Configuração de Interface
   */
  interface: Interface;
  /**
   * Código de Pixel
   */
  pixel: string;
  /**
   * Configuração de Sistema
   */
  system: System;
  [property: string]: any;
}

/**
 * Configuração de Interface
 */
export interface Interface {
  json: Interfacejson;
  text: string;
  [property: string]: any;
}

export interface Interfacejson {
  Layout_Opening: LayoutOpening;
  Sound_Background: SoundBackground;
  Sound_ButtonClick: SoundButtonClick;
  Sound_Loss: SoundLoss;
  Sound_Opening: SoundOpening;
  Sound_Spin: SoundSpin;
  Sound_Win: SoundWin;
  TermsOfService_Text: TermsOfServiceText;
  [property: string]: any;
}
// ...
export interface System {
  json: Systemjson;
  text: string;
  [property: string]: any;
}

export interface Systemjson {
  AdjustmentLimits: AdjustmentLimits;
  BackupEnabled: BackupEnabled;
  BackupInterval: BackupInterval;
  BankHealthDynamicAdjustment: BankHealthDynamicAdjustment;
  BankPercentage: BankPercentage;
  BetValueAdjustment: BetValueAdjustment;
  BreakagePercentage: BreakagePercentage;
  CardPaymentActive: CardPaymentActive;
  ComplianceCheckInterval: ComplianceCheckInterval;
  CPA: CPA;
  CPM: Cpm;
  DepositPercentage: DepositPercentage;
  DisasterRecoveryPlan: DisasterRecoveryPlan;
  DynamicAdjustmentEnabled: DynamicAdjustmentEnabled;
  EnableAuditLogging: EnableAuditLogging;
  EnableSecureBettingProtocol: EnableSecureBettingProtocol;
  EnableServerDivision: EnableServerDivision;
  GameCountAdjustment: GameCountAdjustment;
  GameRoundAdjustment: GameRoundAdjustment;
  LoyaltyPointsRate: LoyaltyPointsRate;
  MaxDeposit: MaxDeposit;
  MaximumBetLimit: MaximumBetLimit;
  MaxRequestsPerMinute: MaxRequestsPerMinute;
  MaxWithdrawal: MaxWithdrawal;
  MinDeposit: MinDeposit;
  MinimumBetLimit: MinimumBetLimit;
  MinWithdrawal: MinWithdrawal;
  ML_DynamicSuccessAnalysis: MLDynamicSuccessAnalysis;
  ML_Enabled: MLEnabled;
  ML_Features: MLFeatures;
  ML_ModelPath: MLModelPath;
  NumberOfMaxUser: NumberOfMaxUser;
  PerformanceMonitoringEnabled: PerformanceMonitoringEnabled;
  PixPaymentActive: PixPaymentActive;
  PlayerSatisfactionResponses: PlayerSatisfactionResponses;
  PlayerSatisfactionSurveyEnabled: PlayerSatisfactionSurveyEnabled;
  PlayerTypeAdjustment: PlayerTypeAdjustment;
  RateLimitingEnabled: RateLimitingEnabled;
  RegistrationStatus: RegistrationStatus;
  RTP_Roulette: RTPRoulette;
  SecureBettingProtocolId: SecureBettingProtocolId;
  SoundEnabled: SoundEnabled;
  StressTestSchedule: StressTestSchedule;
  SurveyFrequency: SurveyFrequency;
  WelcomeBonusPercentage: WelcomeBonusPercentage;
  WinHistoryAdjustment: WinHistoryAdjustment;
  [property: string]: any;
}